+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ObjectTree`: Don't spit out functions and symbols, use a placeholder instead.
 
+### Fixed
+
+- `detailsToggled` event were passed the index as `identifier` instead of the real identifier.
+
 ## [2.1.0] - 2020-05-05
 
 ### Added